+++ HISTORY.md
@@ -5,6 +5,8 @@
   3.x versions.
 - GATK4: Support 4.0 release with changed command line parameters. Re-enable
   multicore calling for CWL runs.
+- install: fix resolution issues due to conda 4.4.x (old ipython-cluster-helper,
+  missing libquadmath.so with numpy due to libgcc update)
 - QC: improve read counting speed by moving to hts-nim-tools, replacing custom
   samtools view counting
 - gVCF: fix basic filtering for GATK and sentieon when running without joint